requirements.txt
appdirs==1.4.4
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
py3dns==3.2.1
pycparser==2.20
pylint==2.6.0
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
validate-email==1.3
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
